Module aptos_metrics_core::op_counters

OpCounters is a collection of convenience methods to add arbitrary counters to modules. For now, it supports Int-Counters, Int-Gauges, and Histogram.

Structs

DurationHistogram

A small wrapper around Histogram to handle the special case of duration buckets. This Histogram will handle the correct granularity for logging time duration in a way that fits the used buckets.

OpMetrics